Consolidate prompt get:
