--- GPTifier/include/prompts.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace prompt
-{
-void get_prompt();
-} // namespace prompt
